# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cellier API",
...
    "contact": {
      "name": "Equipo Cellier"
    },
    "license": {
      "name": "Uso interno"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/households": {
      "get": {
        "tags": [
          "Hogares"
        ],
        "summary": "Listar mis hogares",
        "description": "Los hogares a los que pertenece el usuario autenticado, con su rol en cada\nuno y el número de personas del grupo.\n\nEl orden es estable entre llamadas —por nombre, y por identificador a\nigualdad de nombre—, para que un selector de hogar en el cliente no se\nreordene solo. La lista no incluye el código de ingreso: ese dato sale\núnicamente por el detalle del hogar.\n",
        "operationId": "listMine",
        "responses": {
          "200": {
            "description": "Hogares del usuario. Lista vacía si aún no pertenece a ninguno.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HouseholdSummary"
                  }
                },
                "examples": {
                  "misHogares": {
                    "description": "misHogares",
                    "value": [
                      {
                        "id": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98",
                        "name": "Casa Rivas",
                        "role": "ADMIN",
                        "memberCount": 3
                      },
                      {
                        "id": "b41d0f77-6c58-4e92-8a03-5fd91c2e7a64",
                        "name": "Depa Ñuñoa",
                        "role": "MEMBER",
                        "memberCount": 2
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falta el access token, o no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "sinToken": {
                    "description": "sinToken",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "Se requiere un access token válido.",
                      "instance": "/api/v1/households"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Hogares"
        ],
        "summary": "Crear un hogar",
        "description": "Crea el hogar y deja a quien lo crea dentro como **administrador**, en la\nmisma operación. Un hogar sin administrador sería inadministrable, así que\nesa membresía no es un paso aparte que se pueda olvidar.\n\nEl código de ingreso se genera aquí y viaja en la respuesta.\n",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHouseholdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Hogar creado. La cabecera `Location` apunta a su detalle.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDetail"
                },
                "examples": {
                  "hogarCreado": {
                    "description": "hogarCreado",
                    "value": {
                      "id": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98",
                      "name": "Casa Rivas",
                      "joinCode": "K7M2QP9X",
                      "role": "ADMIN",
                      "memberCount": 1,
                      "createdAt": "2026-09-03T18:42:11Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "El nombre falta, está en blanco o pasa de 80 caracteres.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "nombreVacio": {
                    "description": "nombreVacio",
                    "value": {
                      "type": "https://cellier.app/problems/bad-request",
                      "title": "Datos inválidos",
                      "status": 400,
                      "detail": "Uno o más campos del cuerpo de la petición no son válidos.",
                      "instance": "/api/v1/households",
                      "errors": {
                        "name": "El nombre del hogar es obligatorio"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falta el access token, o no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "sinToken": {
                    "description": "sinToken",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "Se requiere un access token válido.",
                      "instance": "/api/v1/households"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/households/{householdId}/join-code:regenerate": {
      "post": {
        "tags": [
          "Hogares"
        ],
        "summary": "Regenerar el código de ingreso",
        "description": "Emite un código nuevo y **anula el anterior en el acto**: quien lo tuviera\napuntado ya no podrá usarlo para solicitar el ingreso. Es la respuesta a\nque un código se haya difundido más de la cuenta.\n\nLas solicitudes ya enviadas no se ven afectadas: siguen pendientes de que\nun administrador las resuelva. Solo un administrador puede regenerar.\n",
        "operationId": "regenerateJoinCode",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "description": "Identificador del hogar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
          }
        ],
        "responses": {
          "200": {
            "description": "Código regenerado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinCode"
                },
                "examples": {
                  "codigoNuevo": {
                    "description": "codigoNuevo",
                    "value": {
                      "joinCode": "T4WB8HRN"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falta el access token, o no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "sinToken": {
                    "description": "sinToken",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "Se requiere un access token válido.",
                      "instance": "/api/v1/households"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "El usuario es miembro del hogar, pero no administrador.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "noEsAdmin": {
                    "description": "noEsAdmin",
                    "value": {
                      "type": "https://cellier.app/problems/forbidden",
                      "title": "Operación no permitida",
                      "status": 403,
                      "detail": "Esta operación requiere rol de administrador en el hogar.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "El hogar no existe, o existe pero el usuario autenticado no es miembro. La API no distingue ambos casos a propósito.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "hogarAjenoOInexistente": {
                    "description": "hogarAjenoOInexistente",
                    "value": {
                      "type": "https://cellier.app/problems/not-found",
                      "title": "Recurso no encontrado",
                      "status": 404,
                      "detail": "No existe ese hogar, o no eres miembro de él.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/join-requests/mine": {
      "get": {
        "tags": [
          "Hogares"
        ],
        "summary": "Ver mis solicitudes",
        "description": "Todas las solicitudes que ha enviado el usuario autenticado, en cualquier\nestado, de la más reciente a la más antigua.\n\nCada una lleva el nombre del hogar —quien la envió tecleó un código, no un\nnombre— pero ningún otro dato de él: ni su código de ingreso, ni quiénes lo\nhabitan. Eso llega al ser admitido.\n",
        "operationId": "listMine_1",
        "responses": {
          "200": {
            "description": "Solicitudes propias. Lista vacía si no ha enviado ninguna.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyJoinRequest"
                  }
                },
                "examples": {
                  "misSolicitudes": {
                    "description": "misSolicitudes",
                    "value": [
                      {
                        "id": "5a1e93b7-2c04-4d88-91fe-77b6c0a2d431",
                        "householdId": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98",
                        "householdName": "Casa Rivas",
                        "status": "PENDING",
                        "requestedAt": "2026-09-04T09:15:02Z"
                      },
                      {
                        "id": "1b7c40de-9e12-4a63-8f05-33c1a7e6b209",
                        "householdId": "b41d0f77-6c58-4e92-8a03-5fd91c2e7a64",
                        "householdName": "Depa Ñuñoa",
                        "status": "REJECTED",
                        "requestedAt": "2026-08-28T20:41:17Z",
                        "resolvedAt": "2026-08-29T08:03:55Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falta el access token, o no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "sinToken": {
                    "description": "sinToken",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "Se requiere un access token válido.",
                      "instance": "/api/v1/join-requests"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
...
      "CreateHouseholdRequest": {
        "type": "object",
        "description": "Datos para crear un hogar.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del hogar.",
            "example": "Casa Rivas",
            "maxLength": 80,
            "minLength": 0
          }
        },
        "required": [
          "name"
        ]
      },
      "HouseholdDetail": {
        "type": "object",
        "description": "Detalle de un hogar, incluido su código de ingreso.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador del hogar.",
            "example": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
          },
          "name": {
            "type": "string",
            "description": "Nombre del hogar.",
            "example": "Casa Rivas"
          },
          "joinCode": {
            "type": "string",
            "description": "Código para solicitar el ingreso. Conocerlo no da acceso: abre una solicitud\nque un administrador debe aprobar.",
            "example": "K7M2QP9X"
          },
          "role": {
            "type": "string",
            "description": "Rol del usuario autenticado en este hogar.",
            "enum": [
              "ADMIN",
              "MEMBER"
            ],
            "example": "ADMIN"
          },
          "memberCount": {
            "type": "integer",
            "format": "int64",
            "description": "Número de personas que pertenecen al hogar.",
            "example": 3
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación del hogar.",
            "example": "2026-09-03T18:42:11Z"
          }
        }
      },
...
      "HouseholdSummary": {
        "type": "object",
        "description": "Hogar del usuario, con su rol y el tamaño del grupo.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador del hogar.",
            "example": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
          },
          "name": {
            "type": "string",
            "description": "Nombre del hogar.",
            "example": "Casa Rivas"
          },
          "role": {
            "type": "string",
            "description": "Rol del usuario autenticado en este hogar.",
            "enum": [
              "ADMIN",
              "MEMBER"
            ],
            "example": "ADMIN"
          },
          "memberCount": {
            "type": "integer",
            "format": "int64",
            "description": "Número de personas que pertenecen al hogar.",
            "example": 3
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Token JWT emitido por Cellier.",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}